+++ neoscore/interface/graphic_object_interface.py
@@ -1,22 +1,41 @@
+from __future__ import annotations
+
 from dataclasses import dataclass
+from typing import Optional
 
 from neoscore.core.point import Point
 
 
 @dataclass(frozen=True)
-class PositionedObjectInterface:
+class GraphicObjectInterface:
     """Interface for a generic graphic object.
 
     All graphic interfaces for renderable objects should descend from
     this and also be immutable dataclasses.
-
-    ``PositionedObjectInterface`` classes have no concept of parentage, or, by
-    extension, page numbers. Objects creating these interfaces should pass only
-    document-space positions to these.
     """
 
     pos: Point
     """The absolute position of the object in canvas space."""
+
+    parent: Optional[GraphicObjectInterface]
+    """The object's parent, if any.
+
+    If a parent interface is provided, it must be rendered before this interface.
+    """
+
+    scale: float
+    """A scaling factor, where 1 is no scaling."""
+
+    rotation: float
+    """Rotation angle in degrees, where 0 is no rotation."""
+
+    z_index: int
+    """Z-index controlling draw order.
+
+    Use 0 for the default draw order."""
+
+    transform_origin: Point
+    """The origin point for rotation and scaling transforms"""
 
     def render(self):
         """Render the object to the scene.